+++ include/server/server.h
@@ -10,6 +10,8 @@
 #include "error_utils.h"
 #include <atomic>
 #include <vector>
+#include <memory>
+#include "upstream_manager.h"
 #include "tcpworker.h"
 
 using namespace std;
@@ -31,7 +33,8 @@ public:
   Tcpserver(unsigned short port,int backlog,size_t min_threads, size_t max_threads,size_t scale_up_factor , size_t scale_down_factor);
   ~Tcpserver();
 
-  void start(unsigned int num_workers, const Config& config, const std::string& config_path);
+  void start(unsigned int num_workers, const Config& config, const std::string& config_path,
+             std::shared_ptr<UpstreamManager> upstream_manager);
 
   void stop();
 };